requirements.txt
pytest
coverage
pandas
numpycoverage-badge
